Parse Title | Parse body
currently only works with my style of blog entry with php.
needs further fault tolerance and testing to make sure it correctly
parses titles and bodys. right now it will look for the first h2 tag and
parse the inner text as the title. It will spit out the rest into the
body. the body will need further cleaning.